Re-organize the PdbScriptsPipelineStarter file and split it into multiple classes; Change several tutorial
