+++ README
@@ -6,10 +6,16 @@ java: openjdk_1.8.0
 maven: 3.3.9
 mysql: Ver 15.1 Distrib 10.0.21-MariaDB
 blast: 2.4.0+
+*Please make sure java, mvn, mysql, blastp are all in your paths. 
 
 Before you run this project:
-1. Create database schema pdb, username cbio, password cbio in mysql
-2. Download sequences are downloaded from (ftp://ftp.rcsb.org/pub/pdb/derived_data/pdb_seqres.txt.gz) and gunzip it to ${workdir}
+1. Create an empty database schema named as "pdb", username as "cbio", password as "cbio" in mysql:
+	In mysql prompt,type:
+	CREATE USER 'cbio'@'localhost' IDENTIFIED BY 'cbio';
+	GRANT ALL PRIVILEGES ON * . * TO 'newuser'@'localhost';
+	FLUSH PRIVILEGES;
+	create database pdb;
+2. Download PDB sequences from (ftp://ftp.rcsb.org/pub/pdb/derived_data/pdb_seqres.txt.gz) and gunzip it to ${workdir}
 3. Download gene sequences from Ensembl (ftp://ftp.ensembl.org/pub/release-84/fasta/homo_sapiens/pep/Homo_sapiens.GRCh38.pep.all.fa.gz) and gunzip it to ${workdir}
 4. In your code workspace, git clone https://github.com/cBioPortal/pdb-annotation.git
 5. Change settings in org/cbioportal/pdb_annotation/util/Constants.java.
